# docs/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: H-T.Study eLearning Platform API
  version: 1.0.0
  description: |
    API specification cho nền tảng H-T.Study eLearning.
    Hỗ trợ Learning Management, Exam Practice, Gamification, AI Services, Administration.
  contact:
    email: support@htstudy.vn
servers:
  - url: /api/v1
    description: API Version 1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginationMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
        total: { type: integer, example: 150 }
        totalPages: { type: integer, example: 8 }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string, example: "Email is required" }
            timestamp: { type: string, format: date-time }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string, minLength: 3, maxLength: 50 }
        email: { type: string, format: email, maxLength: 254 }
        fullName: { type: string, minLength: 1, maxLength: 100 }
        phone: { type: string, maxLength: 15 }
        avatarUrl: { type: string, format: uri }
        status: { type: string, enum: [active, inactive, locked] }
        roles: { type: array, items: { $ref: '#/components/schemas/RoleSummary' } }
        createdAt: { type: string, format: date-time }
    UserCreate:
      type: object
      required: [username, email, fullName]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, pattern: '^[a-zA-Z0-9_]+$' }
        email: { type: string, format: email }
        fullName: { type: string, minLength: 1, maxLength: 100 }
        phone: { type: string, minLength: 7, maxLength: 15 }
        password: { type: string, minLength: 8 }
        roleIds: { type: array, items: { type: string, format: uuid } }
        groupIds: { type: array, items: { type: string, format: uuid } }
    RoleSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...
    CourseCreate:
      type: object
      required: [title, subjectId]
      properties:
        title: { type: string, maxLength: 300 }
        subjectId: { type: string, format: uuid }
        description: { type: string }
        accessType: { type: string, enum: [free, locked, premium] }
        pointCost: { type: integer, minimum: 0 }
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Initiate SSO/OIDC login
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                redirectUri: { type: string, format: uri }
      responses:
        '200':
          description: Login redirect URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: object, properties: { loginUrl: { type: string } } }
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New tokens issued
...
  /users:
    get:
      tags: [Users]
      summary: List users (Admin)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive, locked] }
        - name: search
          in: query
          schema: { type: string }
        - name: roleId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: User list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/User' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Users]
      summary: Create user (Admin)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          description: User created
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: User details }
    put:
      tags: [Users]
      summary: Update user (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: User updated }
    delete:
      tags: [Users]
      summary: Delete user (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: User deleted }
  /users/{id}/status:
    patch:
      tags: [Users]
      summary: Change user status (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, inactive, locked] }
      responses:
        '200': { description: Status updated }
...
  /courses:
    get:
      tags: [Learning]
      summary: List courses
      parameters:
        - name: subjectId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string, enum: [draft, pending_review, published, rejected] }
        - name: accessType
          in: query
          schema: { type: string, enum: [free, locked, premium] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: Course list with pagination }
    post:
      tags: [Learning]
      summary: Create course (Maker)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseCreate' }
      responses:
        '201': { description: Course created }
  /courses/{id}:
    get:
      tags: [Learning]
      summary: Get course detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Course detail with chapters }
    put:
      tags: [Learning]
      summary: Update course (Maker)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Course updated }
  /courses/{id}/submit:
    post:
      tags: [Learning]
      summary: Submit course for review (Maker)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Submitted for review }
  /courses/{id}/approve:
    post:
      tags: [Learning]
      summary: Approve and publish course (Checker)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Course published }
  /courses/{id}/reject:
    post:
      tags: [Learning]
      summary: Reject course (Checker)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 10 }
      responses:
        '200': { description: Course rejected }

  # ==================== AI SERVICES ====================
...